+++ intergration_test/delete.go
@@ -1,0 +1,10 @@
+package main
+
+// デフォルトで作成されるリクエスト
+// リクエスト:
+//  http.NewRequest("DELETE", "http://localhost:5001", nil)
+// 中身:
+//  DELETE / HTTP/1.1
+//  Host: localhost:5001
+//  User-Agent: Go-http-client/1.1
+//  Accept-Encoding: gzip
