+++ intergration_test/post.go
@@ -1,0 +1,11 @@
+package main
+
+// デフォルトで作成されるリクエスト
+// リクエスト:
+//  http.NewRequest("POST", "http://localhost:5001", nil)
+// 中身:
+//  POST / HTTP/1.1
+//  Host: localhost:5001
+//  User-Agent: Go-http-client/1.1
+//  Content-Length: 0
+//  Accept-Encoding: gzip
